src/requirements.txt
pyyaml==6.0.1
requests==2.31.0
pymongo==4.7.2